ceph: add destination file data sync before doing any remote copy
If we try to copy into a file that was just written, any data that is remote
copied will be overwritten by our buffered writes once they are flushed.  When
this happens, the call to invalidate_inode_pages2_range will also return a
-EBUSY error.

This patch fixes this by also sync'ing the destination file before starting any
copy.

Signed-off-by: Luis Henriques <lhenriques@suse.com>
Reviewed-by: "Yan, Zheng" <zyan@redhat.com>